% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squares.R
\name{area.square}
\alias{area.square}
\title{area.square() function}
\usage{
\method{area}{square}(obj)
}
\value{
The area of a square with \code{side}
}
\description{
area.square() function
}
\examples{
area(square(3))
}

