% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\docType{data}
\name{patients}
\alias{patients}
\title{Toy Patient Data}
\format{
\subsection{\code{who}}{

A data frame with 3 rows and 2 columns:
\describe{
\item{id}{patient id}
\item{age}{patient age}
...
}
}
}
\source{
Lucy's head
}
\usage{
patients
}
\description{
Some fake data on fake patients that I real made up
}
\keyword{datasets}
