% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squares.R
\name{perimeter.square}
\alias{perimeter.square}
\title{perimeter.square() function}
\usage{
\method{perimeter}{square}(obj)
}
\value{
The perimeter of a square with \code{side}
}
\description{
perimeter.square() function
}
\examples{
perimeter(square(3))
}

