% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squares.R
\name{square}
\alias{square}
\title{An S3 class to represent a square
square}
\usage{
square(s)
}
\arguments{
\item{side}{A number to represent the length of the side}
}
\value{
A square object with side value
}
\description{
An S3 class to represent a square
square
}
\examples{
square(3)
}

