\name{square_function}
\alias{square_function}

\title{square_function}

\description{A simple function to calculate the square of a number.}

\usage{
square_function(x)
}

\arguments{
  \item{x}{numeric vector.}
}

\value{The function return a vector.}

\author{Silvan Albin <silvan.albin@hotmail.com>}

\examples{
x <- c(5)
square_function(x)
}