% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squared.R
\name{squared}
\alias{squared}
\title{Squares a number or vector}
\usage{
squared(x)
}
\arguments{
\item{x}{a number to be squared}
}
\value{
the square of a number
}
\description{
Squares a number or vector
}
\examples{
squared(3)
squared(1:3)
}
