% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\docType{data}
\name{patients}
\alias{patients}
\title{World Health Organization TB data}
\format{
\subsection{\code{patients}}{

A data frame with 3 rows and 2 columns:
\describe{
\item{id}{ID}
\item{age}{AGE}
...
}
}
}
\usage{
patients
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
