% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\docType{data}
\name{patients}
\alias{patients}
\title{Toy Patient data}
\format{
\subsection{\code{patients}}{

A data frame with 3 rows and 2 columns:
\describe{
\item{id}{Patient ID}
\item{age}{Patient age}
...
}
}
}
\source{
harrison
}
\usage{
patients
}
\description{
Fake data on patients made up for this package
}
\keyword{datasets}
