% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\docType{data}
\name{patients}
\alias{patients}
\title{Toy patient data}
\format{
\subsection{\code{patients}}{

A data frame with 3 rows and 2 columns:
\describe{
\item{id}{Patient id}
\item{age}{Patient age}
...
}
}
}
\source{
Lucy's head
}
\usage{
patients
}
\description{
Some fake data on fake patients I made up
}
\keyword{datasets}
