% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Se evalua un numero y la funcion botara su cuadrado}
\usage{
square(x)
}
\arguments{
\item{x}{es un numero real}
}
\description{
Se evalua un numero y la funcion botara su cuadrado
}
