% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{A function to square a number}
\usage{
square(x)
}
\arguments{
\item{x}{a number to square (see details).}
}
\value{
y which is the square of x.
}
\description{
Performs the square of a number.
}
\details{
The function computes the square of a number, x must be numerical. Check that x is not a character or factor object.
}
\examples{

x<-2; square(x)
}
\references{
\itemize{
\item{Here you can write a reference related to the function.}
}
}
\author{
\itemize{
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño}
}
}
