% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Squares a number or a vector of numbers}
\usage{
square(x)
}
\arguments{
\item{x}{a number or a numeric vector to be squared}
}
\value{
the squared of x
}
\description{
Squares a number or a vector of numbers
}
\examples{
square(10) # should be 100
square(c(2, 4)) # should be c(4,16)
}
